requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
cryptography==43.0.3
distro==1.9.0
Django==5.1.2
djangorestframework==3.15.2
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
retell-sdk==4.6.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
